+++ src/tui/tui.go
@@ -173,6 +173,8 @@ func (p ColorPair) Bg() Color {
 type ColorTheme struct {
 	Fg           Color
 	Bg           Color
+	PreviewFg    Color
+	PreviewBg    Color
 	DarkBg       Color
 	Gutter       Color
 	Prompt       Color
@@ -272,7 +274,7 @@ type Renderer interface {
 	MaxY() int
 	DoesAutoWrap() bool
 
-	NewWindow(top int, left int, width int, height int, borderStyle BorderStyle) Window
+	NewWindow(top int, left int, width int, height int, preview bool, borderStyle BorderStyle) Window
 }
 
 type Window interface {
@@ -334,12 +336,16 @@ var (
 	ColInfo            ColorPair
 	ColHeader          ColorPair
 	ColBorder          ColorPair
+	ColPreview         ColorPair
+	ColPreviewBorder   ColorPair
 )
 
 func EmptyTheme() *ColorTheme {
 	return &ColorTheme{
 		Fg:           colUndefined,
 		Bg:           colUndefined,
+		PreviewFg:    colUndefined,
+		PreviewBg:    colUndefined,
 		DarkBg:       colUndefined,
 		Gutter:       colUndefined,
 		Prompt:       colUndefined,
@@ -363,6 +369,8 @@ func init() {
 	Default16 = &ColorTheme{
 		Fg:           colDefault,
 		Bg:           colDefault,
+		PreviewFg:    colUndefined,
+		PreviewBg:    colUndefined,
 		DarkBg:       colBlack,
 		Gutter:       colUndefined,
 		Prompt:       colBlue,
@@ -378,6 +386,8 @@ func init() {
 	Dark256 = &ColorTheme{
 		Fg:           colDefault,
 		Bg:           colDefault,
+		PreviewFg:    colUndefined,
+		PreviewBg:    colUndefined,
 		DarkBg:       236,
 		Gutter:       colUndefined,
 		Prompt:       110,
@@ -393,6 +403,8 @@ func init() {
 	Light256 = &ColorTheme{
 		Fg:           colDefault,
 		Bg:           colDefault,
+		PreviewFg:    colUndefined,
+		PreviewBg:    colUndefined,
 		DarkBg:       251,
 		Gutter:       colUndefined,
 		Prompt:       25,
@@ -425,6 +437,8 @@ func initTheme(theme *ColorTheme, baseTheme *ColorTheme, forceBlack bool) {
 	}
 	theme.Fg = o(baseTheme.Fg, theme.Fg)
 	theme.Bg = o(baseTheme.Bg, theme.Bg)
+	theme.PreviewFg = o(theme.Fg, o(baseTheme.PreviewFg, theme.PreviewFg))
+	theme.PreviewBg = o(theme.Bg, o(baseTheme.PreviewBg, theme.PreviewBg))
 	theme.DarkBg = o(baseTheme.DarkBg, theme.DarkBg)
 	theme.Gutter = o(theme.DarkBg, o(baseTheme.Gutter, theme.Gutter))
 	theme.Prompt = o(baseTheme.Prompt, theme.Prompt)
@@ -461,6 +475,8 @@ func initPalette(theme *ColorTheme) {
 		ColInfo = pair(theme.Info, theme.Bg)
 		ColHeader = pair(theme.Header, theme.Bg)
 		ColBorder = pair(theme.Border, theme.Bg)
+		ColPreview = pair(theme.PreviewFg, theme.PreviewBg)
+		ColPreviewBorder = pair(theme.Border, theme.PreviewBg)
 	} else {
 		ColPrompt = pair(colDefault, colDefault)
 		ColNormal = pair(colDefault, colDefault)
@@ -475,6 +491,8 @@ func initPalette(theme *ColorTheme) {
 		ColInfo = pair(colDefault, colDefault)
 		ColHeader = pair(colDefault, colDefault)
 		ColBorder = pair(colDefault, colDefault)
+		ColPreview = pair(colDefault, colDefault)
+		ColPreviewBorder = pair(colDefault, colDefault)
 	}
 }